requirements.txt
langchain 
langchain_community
langchain-qdrant
langchain_groq
langchain_huggingface
langgraph  
fastapi 
streamlit
uvicorn
tavily-python 
beautifulsoup4 
requests
torch --index-url https://download.pytorch.org/whl/cpu
sentence-transformers
qdrant-client
python-dotenv
pydantic
tavily
bs4